+++ backend/controllers/UserController.php
@@ -96,8 +96,7 @@ class UserController extends Controller
                     $model->photo = 'images/' . $imageName . '.' . $model->file->extension;
                     $model->save();
                     $model->file->saveAs('images/'.$imageName .'.'. $model->file->extension);
-                return $this->redirect(['index']); 
-                
+                    return $this->redirect(['index']); 
                 }
                 $model->save();
                 return $this->redirect(['index']); 
@@ -121,35 +120,40 @@ class UserController extends Controller
      */
     public function actionUpdate($id)
     {
-        $session = Yii::$app->session;
-        $model = $this->findModel($id);
-        $session->set('id', $model->idUser);
-        $oldImage = $model->photo;
-        
-        if ($model->load(Yii::$app->request->post())) {  
-            $image = UploadedFile::getInstance($model, 'file');
-            if(UploadedFile::getInstance($model, 'file'))
-            {
-                
-                $model->photo = 'images/' . $model->username .'.'. $image->extension; 
-            }
-            else
-            {
-                $model->photo = $oldImage;
-            }
-            if($model->save())
-            {
-                if(isset($image)){
-                    $image->saveAs($model->photo);   
+        if (\Yii::$app->user->can('atualizarUser')) 
+        {
+            $session = Yii::$app->session;
+            $model = $this->findModel($id);
+            $session->set('id', $model->idUser);
+            $oldImage = $model->photo;
+            
+            if ($model->load(Yii::$app->request->post())) {  
+                $image = UploadedFile::getInstance($model, 'file');
+                if(UploadedFile::getInstance($model, 'file'))
+                {               
+                    $model->photo = 'images/' . $model->username .'.'. $image->extension; 
                 }
+                else
+                {
+                    $model->photo = $oldImage;
+                }
+                if($model->save())
+                {
+                    if(isset($image)){
+                        $image->saveAs($model->photo);   
+                    }
+                }
+                return $this->redirect(['index']);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
             }
-            return $this->redirect(['index']);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
         }
-        
+        else
+        {
+            throw new ForbiddenHttpException;
+        }
     }
 
     /**
@@ -161,36 +165,43 @@ class UserController extends Controller
      */
     public function actionDelete($id)
     {
-        try 
+        if (\Yii::$app->user->can('eliminarUser')) 
         {
-            $model = $this->findModel($id);
-            $books = $model->ideaBook;
-            $addresses = $model->addresses;
-            //testar se o user está relacionado noutra tabela 
-            if($books != null)
+            try 
             {
-                \Yii::$app->session->setFlash('erro', 'Não é possível eliminar este utilizador porque está inserido num ou mais livros de ideias!');
-                return $this->redirect(['index']);
-            }
-            else if($addresses != null)
-            {
-                \Yii::$app->session->setFlash('erro', 'Não é possível eliminar este utilizador porque contém ou mais endereços!');
-                return $this->redirect(['index']);
-            }
-            else
-            {
-                if($model->photo != '')
+                $model = $this->findModel($id);
+                $books = $model->ideaBook;
+                $addresses = $model->addresses;
+                //testar se o user está relacionado noutra tabela 
+                if($books != null)
                 {
-                    unlink(Yii::$app->basePath . '/web/' . $model->photo);
+                    \Yii::$app->session->setFlash('erro', 'Não é possível eliminar este utilizador porque está inserido num ou mais livros de ideias!');
+                    return $this->redirect(['index']);
                 }
-                $this->findModel($id)->delete();        
+                else if($addresses != null)
+                {
+                    \Yii::$app->session->setFlash('erro', 'Não é possível eliminar este utilizador porque contém ou mais endereços!');
+                    return $this->redirect(['index']);
+                }
+                else
+                {
+                    if($model->photo != '')
+                    {
+                        unlink(Yii::$app->basePath . '/web/' . $model->photo);
+                    }
+                    $this->findModel($id)->delete();        
+                    return $this->redirect(['index']);
+                }
+            }
+            catch(\yii\db\IntegrityException $e)
+            {
+                \Yii::$app->session->setFlash('erro', 'Não é possível eliminar este utilizador!');
                 return $this->redirect(['index']);
             }
         }
-        catch(\yii\db\IntegrityException $e)
+        else
         {
-            \Yii::$app->session->setFlash('erro', 'Não é possível eliminar este utilizador!');
-            return $this->redirect(['index']);
+            throw new ForbiddenHttpException;
         }
     }
